+++ views/fabricante_view.py
@@ -1,10 +1,13 @@
 from flask import Blueprint, request, make_response, jsonify, render_template, redirect, url_for
 
 from app import db
-from models import Fabricante, Modelo, Categoria, Marca, Stock
-
+from models import Fabricante
 from schemas import FabricanteSchema
 
+from flask_jwt_extended import(
+    jwt_required,               #para saber si el usuario esta autenticado
+    get_jwt,                    #para saber si el usuario es admin
+)
 fabricante_bp = Blueprint('fabricante', __name__)
 
 # Rutas para Fabricantes
@@ -14,7 +17,13 @@ def listar_fabricantes():
     return render_template('fabricantes.html', fabricantes=fabricantes)
 
 @fabricante_bp.route('/fabricante/nuevo', methods=['GET', 'POST'])
+@jwt_required()
 def nuevo_fabricante():
+    additional_data = get_jwt()
+    administrador = additional_data['administrador']
+
+    if administrador is False:
+        return jsonify(Mensaje="Debes ser admin para poder agregar nuevo fabricante")
     # Código para agregar un nuevo fabricante
     if request.method == 'POST':
         try:
@@ -30,7 +39,13 @@ def nuevo_fabricante():
     return render_template('nuevo_fabricante.html')
 
 @fabricante_bp.route('/fabricante/editar/<int:id>', methods=['GET', 'POST'])
+@jwt_required()
 def editar_fabricante(id):
+    additional_data = get_jwt()
+    administrador = additional_data['administrador']
+
+    if administrador is False:
+        return jsonify(Mensaje="Debes ser admin para poder editar un fabricante")
     # Código para editar un fabricante existente
     fabricante = Fabricante.query.get_or_404(id)
     
@@ -46,7 +61,13 @@ def editar_fabricante(id):
     return render_template('editar_fabricante.html', fabricante=fabricante)
 
 @fabricante_bp.route('/fabricante/eliminar/<int:id>', methods=['POST'])
+@jwt_required()
 def eliminar_fabricante(id):
+    additional_data = get_jwt()
+    administrador = additional_data['administrador']
+
+    if administrador is False:
+        return jsonify(Mensaje="Debes ser admin para poder eliminar un fabricante")
     # Código para eliminar un fabricante
     fabricante = Fabricante.query.get_or_404(id)
     db.session.delete(fabricante)
